+++ R/teal_data_utils.R
@@ -72,5 +72,5 @@ NULL
 #' @rdname teal_data_utilities
 .teal_data_ls <- function(data) {
   checkmate::assert_class(data, "teal_data")
-  grep("._raw_", ls(teal.code::get_env(data), all.names = TRUE), value = TRUE, invert = TRUE)
+  grep("._raw_", ls(teal.code::get_env(data), all.names = FALSE), value = TRUE, invert = TRUE)
 }