+++ NEWS.md
@@ -6,7 +6,8 @@
 * Introduced `teal_transform_module` to provide a way to interactively modify data delivered to `teal_module`'s `server`. #1228
 * Introduced a new argument `once = FALSE` in `teal_data_module` to possibly reload data during a run time.
 * Possibility to download lockfile to restore app session for reproducibility. #479
-* Introduced a function `set_datanames()` to change a `datanames` of the `teal_module`. #1222
+* Introduced a function `set_datanames()` to change a `datanames` of the `teal_module`.
+* Datasets which name starts with `.` are ignored when `module`'s `datanames` is set as `"all"`.
 
 ### Breaking changes